Add multi-result mechanism
If send a huge result iq at one time, the sending may fail.
To solve it, and multi-result mechanism. The result is devided into
many sub-result, the sub-result's status will be 0,1,2... which is
it's order, and last sub-result's staus will be "done:"+Order,
to let server know that it's the last sub-result.

At this time every sub-result's length is about 3000+ byte.