Fix ANR and crash when RemoteManagerService close and restart
Delete the lock when judge mbLogined, because it's useless and
when mbLogined is true, it will not unlock before break;

Add protect in addServerLog, because sometimes it's called when
LogManager is stopped.

Delete the lock in XmppClient.stop, because it's call in main
thread, and handleQuitCmd is called in other thread but also
require the lock, so when thread.join, it will lead to deadlock.

Delete the notifyAll because it must be used in Synchronized.