Get location info asynchronously
Improve asynchronous result mechanism to get location info, and
send all device info to server.

The location provider is not only GPS but all available providers
When 1 of the provide is enabled and callback is not null, the result
will be sent back after get location info.
If no enabled provider or callback is null, send the result directly
with gps="null"
If location callback doesn't come in 60 seconds, also send the result
with gps="null"