Support install and remove across users
Add the support for install/remove apps across users.
When install at first time, install the package for all users, but
block the users who are not request. When install the existing
package for other user, just unblock it.
When remove apps, just block it unless it's current running user.
If all the users block the app, delete it for all.

Note that Our RemoteManagerService is in system server, but
checkCallingPermission will return exception when call some
IPackageManager APIs in the same process. So we use
setApplicationBlockedSettingAsUser which can be used in system
process.
(cherry picked from commit 349520395b0342726f961fbea7b71fafa90e22f0)

Conflicts:

	src/com/zm/epad/plugins/RemotePkgsManager.java