Add FileTransfer, Report Command and RunningApp Result
For the feature that report running app info to server,
we add 2 new class: Command4Report, ResultRunningApp.
Command4Report is the command to require terminal get
Running app ifo, and ResultRunningApp is the result which
will be sent periodically, until untrace command received.

For the feature that set wallpaper with the received image,
we add the class: Command4FileTransfer, which will come with
the URL to download image.

The related definitions in constants are also added.

All above are in zmcommon.

For more detail, please see the XML definition document.