Add IQScheduleResult to send schedule result
Add class IQScheduleResult as the root class to handle schedule
result. There is a HandlerThread in the class. After start, the
thread will send result periodically until stop is called.

If the class is not needed, destroy must be call so that the
thread could end.

By setResultMaker(), you can deside which result to set.

This class can be inherit if needed.