Create RemotePolicyManager
In order to implement device control policy, such as dialy
lock/unlock guest user at a time slot, we create RemotePolicyManager.
This class manage its inner class policy, which is cast to the
policy XML definition. The XML will be delivered by command, and
save as policy.xml. Whenever boot up or policy command is received,
the manager will load all the policies, and set the timers and
policies recorded.
We use cron expression to record the time. So we involved
CronExpression to parse cron time string.
PolicyParser is used to parse the policy XML.

RemotePolicyManager is created when RemoteManagerService start,
then loadPolicy will be called to execute policies.