Create OutputIQCommand to accept policy command
For the commands like policy, the command content must be output
as a file, and which need to be parse is the file. So the handle
process is different from normal command.

We create OutputIQCommand and its provider to handle this kind of
command. The command content will be record as string. After
dispatch, the string will be delivered to the certain manager to
parse or save into a file.