Refactor Command processor
Move RemoteCmdProcessor and ResultFactory from com/zm/epad/plugins
to com/zm/xmpp/communication/handler. Rename RemoteCmdProcessor to
CommandProcessor.

Add class CommandTask to handle command. It represents the handling
task of command. All the command handler will inherit this class.

When a command is received, CommandTask will send EVT_COMMAND to
inform the handler thread to handle the command.
When a result need to be sent, call postResult(), then EVT_RESULT
will be sent to the handler thread with the result.
When a asynchronous task is done, call endTask(), then EVT_TASK_END
will be sent to the handler thread to inform the task is done.

There is also some command which need to pair, e.g. report command.
PairCommandTask is used to handle this kind of command.

For returns
SUCCESS: the task run successfully, and completed when return.
FAILED: the task completely failed.
RUNNING: the task is asynchronous and waiting for callback.