Refactor handing policy command
Delete OutputIQCommand and OutputIQCommandProvider.
Add ZMStringCommand which implements ICommand.

Now all the IQs are parsed by ZMIQCommandProvider. The parseIQ
method in OutputIQCommandProvider is moved to ZMIQCommandProvider,
so that the IQ can be still encapsulated as ZMIQCommand, and
handled by CommandProcessor as common way. In this case, we can
only add CommandTask4Policy to handle policy.

Improved RemotePolicyManager so that when the policy XML is wrong,
a parse exception will be thrown to CommandTask4Policy.Then a NG
result will be sent to server.

Also remove many inactive codes.