Create RemoteAlarmManager
Create RemoteAlarmManager to handle alarm and timer. It can wake up
device no matter whether device is asleep or not.
It uses AlarmManager to set alarm. When alarm go off, a broadcast
will be sent, and received by AlarmReceiver. Then AlarmReceiver
will call back to do something.

What the caller need to do is setAlarm with a callback and a alarm
ID. The ID must be unique in the whole system.