Improve policy and create NotifyListener
Add the handling of AppUsage policy. The app usage report will be
triggered by policy.
Improve starting app policy. It will start the main activity by the
given package name. So add a method in RemotePackageManager to find
the activity.

Create NotifyListener, so that it is possible now to send
information from each manager to outside like CommandProcessor.

The handling in CommandProcessor need to be added.