Fix the bug when set more than 1 alarm
When use Broadcast PendingIntent to set alarm, if the intent action
is the same, it will get the same pending intent. So if set alarm
with the same action, the alarm will be just updated instead of set
2 different alarms.

So fix it by using different intent action to set alarm. Now there
will be many actions prepared to be used by alarm. All these
actions are registered to the receiver.When set alarm,it will get
an action which is not used, and set it as the intent to broadcast
when alarm go off.