Create Accumulate Policy and implement protect eye
Accumulate Policy is created to implement the policy that after a
time accumulation, trigger a certain action.

Protect eye policy is implemented by Accumulate Policy. The time
accumulation is trigger by receive system screen on/off broadcast.
When time out, it will call for disable screen.

A few new feature in RemoteDeviceManager need to be added.
Now disableScreen() is just a stub