Upload log files by command
The command of uploading log files will be sent as Command4Query,
and handled by CommandTask4Query.

After upload log file, if update flag is true, delete current file.
But when upload today's log, because the log files are open, and
keep writing, we must close it. There is also a problem of name
duplication when upload today's log many times, becase the name in
one day is always the same. So we do as below:
1, Close current log file
2, Rename current log file
3, Create a new empty log file
4, Upload the renamed log file
5, After upload succeed, delete the renamed log file

If it's not today's log, no need to rename because it won't be
written any more, and after delete no more log file need to created