Save result to log file when error IQ is received
When error IQ happens, NetCmdDispatcher will receive it, and notify
the listener via new method handleError(). CommandProcessor as a
listener will save the error in the list. After sendPacket, it will
check the result again after 2 second. If there is an error for the
result, the result will be saved in log file.

Also use the same iq id when send iq result.