Adjust zm common constants
Coordinate with server team, and decide to adjust the Constants in
zm common lib. It impact many string definition for both command
and result.

Now all of the commands have their own unique action. When send
normal result, the result must be with this action so that server
could know what the origin action is. The most change of it is
report command, report command action is changed, and report tag
is not used any more. Running App report is not supported any more,
the feature is moved to query command which will only send result
once.

The normal result status is not "OK"/"NG" any more. The right
String is defined as RESULT_OK, RESULT_ERR, and RESULT_Done in
Constants file.

Modify debug UI for above changes.

Also recover icon in laucher for easy debug.