Open API via AIDL and complete login feature
Create AIDL API, and encapsulate it into RemoteManager class. The
API classes are built into a jar. Other apps can use our API by
importing this jar.

Now there is one API: login. This API is used for login at the
first time when user register his/her terminal. If the username
and password are right, the info will be saved in config file, and
auto-login in next time. Otherwise, the login API will return
false. This is a synchronious API, we encapsulate an asynchronious
execution into synchronious method. There is also modification to
handle the cases about wrong login info, as well as non-login
status. If there is no username in config, it means the ternimal
has not loginned, in this case, don't reconnect.

Devide XmppClient.start method into 2 methods: start and connect.
start() is used to create the XmppClient handle thread, and
connect() is used to connect to XMPP server. This change is
because we always need to keep the handle thread, but connect is
call time by time. For example, if the password is wrong, we have
to connect again otherwise authentication exeption will always
happen.

Also delete classpath from git because every developers' classpath
is different.
Also change log tag from com.zm.epad:TAG to com.zm.epad.TAG,
because ":" is difficult to use when filter via adb logcat.