Keep wake up when reconnect to XMPP server
When device is sleeping, if try to connect and login, it probably
goes to sleep again connect or login, so that connection can't
recovery. So we have to keep device awake when reconnect.

Add PowerManager.WakeLock acquire/release in RemoteDeviceManager.
Also add the API in SubSystemFacade, the default wake level is
PARTIAL_WAKE_LOCK.

The WakeLock APIs are called when reconnect.

Also clean RemoteDeviceManager.jave format.