Use PingManager to handle XMPP ping
There is always ping sent from server to check if client is
availble, if don't response with pong, the connection will be
closed. So use PingManager to handle the ping to keep the
connection stable.

When reconnect, if the connection exists, no need to create a new
connection or PingManager.