Improve package download and install process
When receive install command, firstly check if the package exists
and the version is same. If package doesn't exist or version
update, download the new apk and install it. Otherwise, intall the
existed package, or send "already existed" error.

The app result is not normal any more, we must give package info
like package name, version, to the server, so that the server can
know the action.

Change the return value of install to int, so that we can get more
info when install fails, and send it as error code.