Use Wifi MAC address as XMPP username
Change login info and use wifi MAC address as XMPP username. It
provides an unique username which is fixed when produce in factory
and avoid dulplicated name. Now when login to XMPP server, the only
info which need to input is the password. Also when check if the
device account is initiated, only check if the password is set.

Resource is changed to "default" as the request of server team.

Change login API to adapt above change.