Handle asynchronous commands
This feature is to handle asynchronous commands, which are sent
when device is offline.

The program of synchronous commands is triggered after Xmpp login.
CommandProcessor listens the login event and ask to obtain commands
obtained from WebService. the URL is saved in Config.

After obtain the commands, ZMIQCommandProvider is used to parse the
commands. The new method parseCommands is used to do this. After
parser, the commands will be posted by postCommand.